Feat/turn off redux logger in prod and clean up routing (#19)
* Starting point new api delegations

* add react-redux

* Added icon

* backup

* backup

* backup

* fix rtk for ts

* use redux in new api delegations Page

* lint fix

* change naming

* yarn i

* Clean up

* refactor code

* Cleanup

* Design system update

* Code cleanup

* first iteration

* Finish implementation, turn off redux logger in prod

* Cleanup router
