--- src/routes/ApiDelegationOverview.tsx
@@ -1,5 +1,0 @@
-import { ApiDelegationOverviewPage } from '../components/ApiDelegationOverviewPage/ApiDelegationOverviewPage';
-
-export const ApiDelegationOverview = () => {
-  return <ApiDelegationOverviewPage />;
-};
