+++ src/routes/RouterElements/ApiDelegationOverview.tsx
@@ -1,0 +1,5 @@
+import { ApiDelegationOverviewPage } from '../../components/ApiDelegationOverviewPage/ApiDelegationOverviewPage';
+
+export const ApiDelegationOverview = () => {
+  return <ApiDelegationOverviewPage />;
+};
