+++ src/routes/RouterElements/NewApiDelegations.tsx
@@ -1,4 +1,4 @@
-import { NewApiDelegationsPage } from '../components/NewApiDelegationPage';
+import { NewApiDelegationsPage } from '@/components/NewApiDelegationPage';
 
 export const NewApiDelegations = () => {
   return <NewApiDelegationsPage />;
